requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.0
sqlparse==0.5.1
tzdata==2024.2
